hw01: Vectors, strings and functions
Task:

We are given a file of text that has been encrypted. For simplicity, we will assume the name is "encrypted.txt".  That will save you having to ask the user.
Do not ask the user for the file name.
Luckily we know what algorithm was used, the classic Caesar cypher. Each UPPERcase alphabetic character has been modified by “rotating” it through the alphabet some number of steps.
For example, if the number of steps was 3 and the original letter was 'B' then the encrypted file would instead have a 'E' in the same place, since 'E' is three letters past 'B' in the alphabet. On the other hand, if the rotation distance was -3 and the letter was 'E' in the encrypted file, then the unencrypted character would be 'B'.
Clearly we have to address the problem of what to do when the letter to be encrypted is too close to the end of the alphabet. For example, what should we do if we had to encrypt the letter 'Y', again with the number of steps specified as three. The first step takes us to 'Z'. What about the second step? We wrap around, back to 'A'. The third and final step then is easy, it will take us on to 'B'.
Of course, we are decrypting, so the task for our program is to realize that if we see a 'B', then it was originally a 'Y'.
Your program will read an encrypted file. You are to decrypt it.
The first item in the file will be an int, the number of steps that was used in the rotation. Yes, it could be negative.
The following lines will contain the text that you are to decode. Only UPPERcase alphabetic characters have been encrypted. The rest, e.g. lowercase, digits and punctuation all appear in the encrypted file the same way they did in the original.
Oh, one other thing. To make the encoded file just a little more difficult, they thought it would be clever to reverse the lines in the file. So what should be the first line shows up last, etc.
You will be reading the lines from the file into a vector of strings, in the order that you read them. Of course the number of items in your vector, i.e. its size, will be the same as the number of text lines in the input file.
Do not reorder the lines in the vector!
When you are displaying the decoded lines, you will just loop through the vector backwards. (An example of doing this will have been covered in class.)
An example. Suppose this is the contents of the file you were asked to decrypt:
5
FSI YMJ RTRJWFYMJX TZYLWFGJ.
FQQ RNRXD BJWJ YMJ GTWTLWTAJX
INI LDWJ FSI LDRGQJ NS YMJ BFGJ.
YBFX GWNQQNL FSI YMJ XQNYMJD YTAJX
That says that the original file was encoded with a Caesar cypher using a rotation of five. Remember that the lines have also been reversed meaning that the first line of text we see is the encoded version of what was the last line of code in the original file, etc. When we decrypt it, we should get:
TWAS BRILLIG AND THE SLITHEY TOVES
DID GYRE AND GYMBLE IN THE WABE.
ALL MIMSY WERE THE BOROGROVES
AND THE MOMERATHES OUTGRABE.


Overall Feedback
Deductions:
-1: must close the stream
-1: did not get rid of newline after the int is read (debug program to see that index 0 contains "")
-1: magic number, should be using characters in decryptChar instead of ASCII values ("A" instead of 64)


Warnings:
- filename should be hw01.cpp not hw1.cpp
- ensure that comments on function definitions include (1. Description, 2. Parameters, 3. Return Type)
- lines must be less than 80 characters long

-AA
Score
17 / 20 - B