hw06: Operator Overloading
Focus

Operator Overloading
Task

You will implement a class called BigUnsigned. This class will represent unsigned integers that can have arbitrarily many digits. In C++, the type unsigned is normally limited to non-negative values less than 232. You will use a vector of ints, with each entry in the vector representing a single digit. Store the digits with the low order digit first, i.e. for the number 123, index 0 of the vector would hold the value 3. This will simplify some of the functions you have to write.

You will provide:

Constructors:
One that can take an int, defaulting to the value 0 if no argument is provided.
One that takes a string. The string is intended to be a string of digits.
To allow for possible errors, if the string does not start with any digits, then use the value 0.
If it starts with a digit, but has other characters, then just use the digits up until the first non-digit character.
If there additionally "leading" zeros, they should be skipped or removed.
Operators
Output
Addition
Plus-equals
Increment operators, both pre- and post-.
Comparison operators: equals, not-equals, less than, less than or equals, greater than, greater than or equals.
Follow the conventions that we have discussed to decide whether to implement each operator as a member or non-member function.
Maximize your code reuse so that it is not necessary to repeat yourself. 
Be sure that the test code that we provide works and will generate the correct output.
Maybe we used an operator that wasn't mentioned above?
Be sure to add it!
Attached is a test program and corresponding output.


Overall Feedback
Errors:
-1 only checked the first spot for carry over, not every spot with op++. This would cause an error when doing an operation with op++ on a number like 99, should be using the op+= method in op++

-AA
Score
19 / 20 - A