rec05: Classes, delegation, output operator, copy control
Topics

Design interacting classes using
Association
Composition
Delegation
Copy control: destructor and copy constructor.
Nesting classes
Output operators
In this lab you will be modeling a little part of the School of Engineering CS lab administration.

You are being asked to write software so that the University programmers can model the way your lab instructors keep up with your grades.

We've reduced this problem a great deal so that there is not too much for you to do.
The complete set of code would be really long and would require days or weeks of analysis.

Remember that the focus of the lab is to practice encapsulation, data hiding, association and delegation. The first two are basic tenets of OOP. Delegation means that if one object holds a piece of data and you want something done with that data, then you ask the object to do it for you.