rec07: Operator overloading
Topics

Overloading Operators
Recitation Project

Design, implement and test a C++ class to work with real numbers as rational numbers.
The data members are the numerator and denominator, stored as integers.
We have no need for setters or getters (aka mutators or accessors) in the Rational class. All of our operations on rational numbers involve the entire number, not just the numerator or denominator.
Implement the operators:
<< and >> (i.e., input and output).
Rational numbers are read and written as
an integer,
followed by a slash,
followed by an integer.
Either (or both) of the numerator and the denominator may be input as negative integers. The following are all possible inputs: 1/2, -1/-2, -1/2 and 1/-2.
It's a design decision as to when to "normalize" a rational number. Should it always be normalized? Or is that only a choice to be made when we display it?
We will maintain the numbers in their normalized form at all times.
Thus when they are first created, they must be stored as normalized and any operation that might change a number must also normalize it.
What does "normalize" mean? The numerator and denominator will be in "lowest terms" and only the numerator may be negative. For example, if we were given as input 4/-8, we would store and display it as -1/2. Similarly -8/-6 would turn into 4/3.
+= Implement as a member function, aka method.
+ Implement (i.e., addition) as a non-member function that calls the += operator. Do not make + a friend. (There is no reason to.)
== and <: Implement as non-member and as friend.
++ and --
Both pre- and post-.
Member for ++
Non-member, non-friend for --
Yes, normally we would do this as a member.
We are just asking you to implement the decrement operator as a non-member for you to see how it would look different.
!=, <=, > and >= Implement as non-member and non-friend.
Make it possible to write if (r) {}, where r is a Rational number. The test will evaluate to false if the numerator is zero and true otherwise.
We provide:
rec07_begin.cpp to test your class. Can you think of additional tests to write?
It contains a definition for the function greatestCommonDivisor.
Sample output, where I used 4/8 and 1/-2 for my input values.